'Lazy' one-to-many collection

+++ lib/clipper/mapping/associations/one_to_many.rb
@@ -3,26 +3,58 @@ module Clipper
 
     class OneToManyCollection < ::Clipper::Collection
 
-      def initialize(association, parent, children)
+      def initialize(association, parent, children = nil)
+        @loaded = !children.nil?
+        children = [] if children.nil?
+        
         @association = association
         @parent = parent
         @collection = children
         @mapping = association.associated_mapping
       end
 
+      def loaded?
+        @loaded
+      end
+
       def add(item)
+        load! unless loaded?
         super
 
-        @parent.__session__.enlist(item) if @parent.__session__
-        @association.set_key(@parent, item)
+        if @parent.__session__
+          @association.set_key(@parent, item)
+          @parent.__session__.enlist(item)
+        end
 
         item
       end
       alias << add
 
+      def each
+        load! unless loaded?
+        @collection.each { |item| yield item }
+      end
+
+      def size
+        load! unless loaded?
+        @collection.size
+      end
+
+      protected
+
+      def load!
+        if @parent.__session__
+          criteria = @association.match_criteria.call(@parent, Clipper::Query::Criteria.new(@mapping))
+
+          @collection = @parent.__session__.find(@mapping, criteria.__options__, criteria.__conditions__)
+        end
+
+        @loaded = true
+      end
     end
 
     class OneToMany < Association
+      attr_reader :match_criteria
 
       def initialize(mapping, name, mapped_name, &match_criteria)
         raise ArgumentError.new("You must pass a block containing a criteria expression for '#{mapping.name} has_many #{name}'") unless match_criteria
@@ -74,36 +106,30 @@ module Clipper
           if data = instance_variable_get(association.instance_variable_name)
             data
           else
-            if __session__
-              instance_variable_set(association.instance_variable_name, OneToManyCollection.new(association, self, association.load(self)))
-            else
-              instance_variable_set(association.instance_variable_name, OneToManyCollection.new(association, self, [])) #Collection.new(association.associated_mapping, [])))
-            end
+            instance_variable_set(association.instance_variable_name, OneToManyCollection.new(association, self))
           end
         end
 
         target.send(:define_method, association.setter) do |new_value|
           raise ArgumentError.new("#{self.class}.#{association.setter} only accepts enumerables") unless new_value.is_a?(Enumerable)
 
-          if items = self.send(association.getter)
-            if __session__
-              items.each do |item|
-                association.unlink(self, item)
-                __session__.enlist(item)
-              end
+          if (items = self.send(association.getter)) and __session__
+            items.each do |item|
+              association.unlink(self, item)
+              __session__.enlist(item)
             end
           end
 
-          new_value.each do |item|
-            association.set_key(self, item)
-            __session__.enlist(item) if self.__session__
-#            puts self.send(association.getter).size if self.__session__
+          if self.__session__
+            new_value.each do |item|
+              association.set_key(self, item)
+              __session__.enlist(item)
+            end
           end
 
           instance_variable_set(association.instance_variable_name, OneToManyCollection.new(association, self, new_value))
         end
       end
     end
-
   end
 end
